moved uxpProps to separate folder
